Restrucutre of gate enum
The gate enum and it's private structs such as GateInfo and GateSize
have been isolated to their own filed, as a sub-module of circuit.
Again, this will hopefully increase the readability and development ease
of tthe code.
