+++ examples/custom_gate.rs
@@ -43,7 +43,7 @@ fn main() -> Result<(), QuantrError> {
 
 // Implements the CCC-not gate.
 fn cccnot(input_state: ProductState) -> SuperPosition {
-    let state: Vec<Qubit> = input_state.qubits;
+    let state: Vec<Qubit> = input_state.qubits.clone();
     let state_slice: [Qubit; 4] = [state[0], state[1], state[2], state[3]]; // In this format, this
                                                                             // guarantees that state_slice has length 4 to the rust compiler. Useful for the match
                                                                             // statement.
@@ -56,8 +56,6 @@ fn cccnot(input_state: ProductState) -> SuperPosition {
                 .unwrap()
                 .into_super_position()
         }
-        other_state => ProductState::new(&other_state)
-            .unwrap()
-            .into_super_position(),
+        _ => input_state.into_super_position(),
     }
 }
