+++ src/circuit/states/qubit.rs
@@ -41,4 +41,3 @@ impl Qubit {
         ProductState::new_unchecked(&[self])
     }
 }
-
