+++ src/lib.rs
@@ -15,8 +15,9 @@ mod complex;
 mod error;
 
 // Make available for public use.
+pub use circuit::gate::Gate;
 pub use circuit::printer::Printer;
 pub use circuit::states;
-pub use circuit::{Circuit, Gate, Measurement};
+pub use circuit::{Circuit, Measurement};
 pub use complex::{Complex, COMPLEX_ZERO};
 pub use error::QuantrError;
